# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ThingTypes API
    version: 0.0.1
paths:
    /model-manager/v1/thingTypes:
        post:
            tags:
                - ThingTypes
            operationId: ThingTypes_CreateThingTypes
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateThingTypesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateThingTypesReply'
                default:
                    description: Default errors response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        Characteristic:
            type: object
            properties:
                name:
                    type: string
                unit:
                    type: string
                length:
                    type: string
                dataType:
                    type: string
                defaultValue:
                    type: string
        CreateThingTypesReply:
            type: object
            properties:
                name:
                    type: string
                tenant:
                    type: string
                parentTypeId:
                    type: string
                description:
                    type: string
                characteristics:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/Characteristic'
                propertySets:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/PropertySet'
                meta:
                    $ref: '#/components/schemas/Meta'
        CreateThingTypesRequest:
            type: object
            properties:
                name:
                    type: string
                parentTypeId:
                    type: string
                description:
                    type: string
                characteristics:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/Characteristic'
                propertySets:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/PropertySet'
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        Meta:
            type: object
            properties:
                id:
                    type: string
                version:
                    type: string
                tenant:
                    type: string
                createdById:
                    type: string
                updatedById:
                    type: string
                createdByName:
                    type: string
                updatedByName:
                    type: string
                createdTime:
                    type: string
                    format: date-time
                updatedTime:
                    type: string
                    format: date-time
        Property:
            type: object
            properties:
                name:
                    type: string
                unit:
                    type: string
                value:
                    type: string
                dataType:
                    type: integer
                    format: enum
                accessMode:
                    type: integer
                    format: enum
                min:
                    type: string
                max:
                    type: string
        PropertySet:
            type: object
            properties:
                properties:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/Property'
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing errors message, which should be in English. Any user-facing errors message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the errors details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical errors model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: errors code, errors message, and errors details. You can find out more about this errors model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: ThingTypes